refactor(index): split entrypoint into app/server/services
Break src/index.ts (570 lines) into focused modules:

- app.ts: Express app factory (middleware, routes, probes, error handler, 404)
- server.ts: HTTP server creation + WebSocket attachment
- services.ts: background-service startup orchestration

index.ts stays a thin entrypoint: tracer init, wiring, listen, and graceful
shutdown. Also fixes a duplicate httpServer creation and calls to unimported
WS attach helpers that were present in the original file.
